+++ ChangeLog.md
@@ -3,11 +3,14 @@
 * Multi-measure stamp tracking much better.
 * Bug fix to new stamp/instrument->score assignment.
 * Rippling shaders for dancing robit icons
+* BallWorld drawing optimization
+* Rainbow shader stood up in app
 
 11/21/2016 - 11/26/2016
 * Multi-measure tracks (and short tracks) works properly now.
 * New stamp system for instrument assignment.
 * More awesome character animation.
+* New Sharp Crab character
 
 11/14/2016 - 11/18/2016
 * Barrel distortion correction!